+++ tests/fixtures/UserFixture.php
@@ -2,9 +2,9 @@
 
 declare(strict_types=1);
 
-namespace davidhirtz\yii2\config\tests\fixtures;
+namespace Hirtz\Config\tests\fixtures;
 
-use davidhirtz\yii2\skeleton\models\User;
+use Hirtz\Skeleton\models\User;
 use yii\test\ActiveFixture;
 
 class UserFixture extends ActiveFixture