+++ tests/functional/AuthCest.php
@@ -10,15 +10,15 @@ declare(strict_types=1);
  * @noinspection PhpUnused
  */
 
-namespace davidhirtz\yii2\config\tests\functional;
+namespace Hirtz\Config\tests\functional;
 
-use davidhirtz\yii2\config\modules\admin\models\Config;
-use davidhirtz\yii2\config\modules\admin\widgets\forms\ConfigActiveForm;
-use davidhirtz\yii2\config\tests\support\FunctionalTester;
-use davidhirtz\yii2\skeleton\codeception\fixtures\UserFixtureTrait;
-use davidhirtz\yii2\skeleton\codeception\functional\BaseCest;
-use davidhirtz\yii2\skeleton\models\User;
-use davidhirtz\yii2\skeleton\modules\admin\widgets\forms\LoginActiveForm;
+use Hirtz\Config\modules\admin\models\Config;
+use Hirtz\Config\modules\admin\widgets\forms\ConfigActiveForm;
+use Hirtz\Config\tests\support\FunctionalTester;
+use Hirtz\Skeleton\codeception\fixtures\UserFixtureTrait;
+use Hirtz\Skeleton\codeception\functional\BaseCest;
+use Hirtz\Skeleton\models\User;
+use Hirtz\Skeleton\modules\admin\widgets\forms\LoginActiveForm;
 use Yii;
 
 class AuthCest extends BaseCest