+++ tests/support/FunctionalTester.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace davidhirtz\yii2\config\tests\support;
+namespace Hirtz\Config\tests\support;
 
 /**
  * Inherited Methods